changed roomba_sevenseg_digits to int32_t

+++ include/program.h
@@ -60,7 +60,7 @@ void program_run();
   * \return	1 if the integer value is smaller than 10000, -1 otherwise
   */
 
-int8_t intToAscii(int32_t value, uint8_t out[]);
+int8_t intToAscii(int32_t value, int32_t out[]);
